moved test to main
